% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to a database}
\usage{
connect(dbName = "NA")
}
\arguments{
\item{dbName}{Name of connection which needs to be connected. Use list_connection() in case you are not sure about the connection name.}
}
\description{
This function allows you to see all the saved connections of different DBs.
}
\examples{
conn <- connect(dbName)
}
