% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_connection.R
\name{create_connection}
\alias{create_connection}
\title{Create Connection}
\usage{
create_connection()
}
\description{
This function allows you to create connections to different DBs.
}
\examples{
create_connection()
}
