% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_email_profile.R
\name{create_email_profile}
\alias{create_email_profile}
\title{Create/Update Email connection}
\usage{
create_email_profile()
}
\description{
This function allows you to create/update email profiles to send out emails.
}
\examples{
create_email_profile()
}
