% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_connection.R
\name{delete_connection}
\alias{delete_connection}
\title{Delete existing connections}
\usage{
delete_connection(Connection_Name = "NA")
}
\arguments{
\item{Connection_Name}{Name of connection which needs to be deleted. Use list_connection() in case you are not sure about the connection name.}
}
\description{
This function allows you to delete a saved connections.
}
\examples{
delete_connection(Connection_Name="")
}
