% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_query.R
\name{execute_query}
\alias{execute_query}
\title{Executes a Query}
\usage{
execute_query(connection_name = "ADPG", Query = "NA", prefix = "NA",
  split = FALSE, retry = 0, sleep = 0, logs = getwd())
}
\arguments{
\item{connection_name}{Name of connection on which the query needs to be executed}

\item{Query}{Complete query to be executed}

\item{prefix}{Prefix to be added before every temp table}

\item{split}{Set it to TRUE if one needs to execute queries one by one. Split happens using the delimiter ()}

\item{retry}{Number of retry attempts on failure}

\item{sleep}{Number of seconds to sleep before retrying}

\item{logs}{No need to provide}
}
\description{
This function allows you to pass a complete code where individual queries are seperated by ';'. It returns a dataframe with whether the code was successfully executed or not. Should only be used when the queries start with Create, insert or drop commands.
}
\examples{
df <- execute_query(connection_name,Query)
}
