% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data into R from a DB}
\usage{
fetch_data(connection_name = "NA", Query = "NA", bulk_load = 0,
  memory = 8)
}
\arguments{
\item{connection_name}{Name of connection on which the query needs to be executed}

\item{Query}{Complete query to be executed}

\item{bulk_load}{(Optional) boolean field to be changed to one only if volume of data is high}

\item{memory}{(Optional) Java Memory in GB that needs to be set if bulk_load fails. Max Value 20GB}
}
\description{
This function allows you to Fetch data from a source.
}
\examples{
df <- fetch_data(connection_name="",Query="",..)
}
