% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{get user info for a connection}
\usage{
get_user(dbName = "NA")
}
\arguments{
\item{dbName}{Name of connection which needs to be connected. Use list_connection() in case you are not sure about the connection name.}
}
\description{
This function allows you to get the user name for a saved connection!
}
\examples{
user <- get_user(dbName)
}
