% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_connections.R
\name{list_connections}
\alias{list_connections}
\title{List existing connections}
\usage{
list_connections()
}
\description{
This function allows you to see all the saved connections of different DBs.
}
\examples{
df <- list_connections()
}
