% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail}
\alias{send_mail}
\title{Send emails from R}
\usage{
send_mail(from, to, cc = NULL, bcc = NULL, subject = "", body = " ",
  attachments = NULL, html = FALSE, inline = FALSE)
}
\arguments{
\item{from}{A valid email address of the sender.}

\item{to}{A character vector of recipient valid email addresses.}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email. For Adding tables and graphs (see details for more info)}

\item{attachments}{A character vector of paths in the file system linking to files or *valid* URLs to be attached to the email (see details for more info on attaching URLs)}

\item{html}{Default value False. Change it to true if there are tables or graphs to be inserted.}

\item{inline}{Default value False. Change it to true if there are tables or graphs to be inserted.}
}
\description{
This function allows you to send emails with attachments. One can use mailR library as well.
}
\details{
Using 'attachments' you can attach files or webpages hosted on the web (for e.g. on Dropbox). Currently, URLs hostnames must be prepending with http:// or https://.
}
\examples{
sender <- "sender@gmail.com"  # Replace with a valid address
recipients <- c("receiver1@gmail.com")  # Replace with one or more valid addresses
attachments <- c("C:/file_name.txt","https://gmail.com")
send_mail(from = sender,
         to = recipients,
         subject="Subject of the email",
         body = "Body of the email",
         attachment = attachments)

library(xtable)
file_name <- "C:/Users/Admin/Documents/sample_data.csv"
sample_data <- read.csv(file_name)
plot_file <- "sample_plot.png"
## Converts table in html format to be attached inside mail body
sample_table <- print(xtable(dataframe,caption = "Heading for the table"), type="html", caption.placement = "top")
## Set sender and recepient
sender <- "qv_mailer@swiggy.in"
recipients <- c("to_list_1@swiggy.in","to_list_2@swiggy.in")
## Export sample plot to image file
png(filename=plot_file)
plot(sample_data)
dev.off()
## Set subject of the mail
subject <- "Subject Line"
## Mail body opening
mail_body <- paste0("greetings here, </b><br><br> additional comments here")
## Attach table
mail_body <- paste0(mail_body,"</b><br><br>",sample_table)
## Embed charts
mail_body <- paste0(mail_body,'</b><br> Sample plot: </b><br><img src="sample_plot.png">')
 ## Adding P.S. and signature
mail_body <- paste0(mail_body,"</b><br>more comments here.<br><br>signature here")
}
